+++ src/TTX.App/Dto/Creators/VoteDto.cs
@@ -6,7 +6,7 @@ namespace TTX.App.Dto.Creators;
 public class VoteDto
 {
     [JsonPropertyName("creator_id")] public required int CreatorId { get; init; }
-    [JsonPropertyName("value")] public required long Value { get; init; }
+    [JsonPropertyName("value")] public required double Value { get; init; }
     [JsonPropertyName("time")] public required DateTimeOffset Time { get; init; }
 
     public static VoteDto Create(Vote vote)
